requirements.txt
numpy==1.21.0
pandas==1.2.5
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
